Add vibe coded counter
Created by Gemini 2.5 Pro preview

With 2 prompts:
1/
```
Create a single HTML page with Javascript and CSS definitions.
The page must contain a full screen timer, counting down to Sept 4 2025. Make the counter the only eye-catching element on the page, attracting all the attention but also make the site visually appealing.
The website must be non-scrollable
Make the site beautiful. The color palette should be build around #3B1D62 and #880c51 and white fonts - mimicking the enclosed screenshot of a website
```

2/
```
- make it use tailwindCSS, imported from a public CDN
- put minutes and seconds in the second row for desktop and tablet resolutions as well
```